rename rest of the package
